refactor test to be more high level
