+++ Tests/PlayMode/SwitchViewModeTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3c9c827a726c3b94488aeb85d80ea534
+guid: 6c881a48511cdab41aebe285f945dc4b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
